refactor: Participant 에서만 사용되는 Money 를 participant 패키지로 이동
